* Fix potential ArgumentError from OptionParser
OptionParser was, at times, interpreting the block passed to
BinDiesel's opts_on as an additional argument for OptionParser#on,
which caused OptionParser#notwice to throw an ArgumentError (the
block was interpreted by OptionParser as a proc passed to the switch,
and BinDiesel was also passing in the block as a separate argument,
so it ended up being passed in twice).

To fix this, we are removing any Proc parameters from
BinDiesel#opts_on (the only feasible reason to pass a Proc to
BinDiesel#opts_on would be to handle the value from the parsed option,
which is generally handled by the block provided to BinDiesel#opts_on).